+++ src/test/java/com/map/gaja/client/presentation/api/ClientPutControllerTest.java
@@ -5,13 +5,11 @@ import com.map.gaja.client.apllication.ClientAccessVerifyService;
 import com.map.gaja.client.apllication.ClientQueryService;
 import com.map.gaja.client.apllication.ClientService;
 import com.map.gaja.client.domain.exception.ClientNotFoundException;
-import com.map.gaja.client.infrastructure.file.FileValidator;
 import com.map.gaja.client.infrastructure.s3.S3FileService;
-import com.map.gaja.client.presentation.ClientRequestValidator;
+import com.map.gaja.client.infrastructure.ClientRequestValidator;
 import com.map.gaja.client.presentation.dto.request.NewClientRequest;
 import com.map.gaja.client.presentation.dto.subdto.StoredFileDto;
 import com.map.gaja.group.application.GroupAccessVerifyService;
-import com.map.gaja.group.domain.exception.GroupNotFoundException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
